+++ tests/TestCase/Http/Controller/ControllerTest.php
@@ -442,7 +442,7 @@ class ControllerTest extends \PHPUnit\Framework\TestCase
         $book->name = 'How to use PHPUnit';
         $controller->set('book', $book);
         $controller->serialize('book');
-        $request->header('Accept', 'application/json');
+        $request->headers->set('Accept', 'application/json');
         $controller->render();
         $this->assertEquals($book->toJson(), $controller->response()->body());
     }
@@ -482,7 +482,7 @@ class ControllerTest extends \PHPUnit\Framework\TestCase
         $book->name = 'How to use PHPUnit';
         $controller->set('book', $book);
         $controller->serialize('book');
-        $request->header('Accept', 'application/xml');
+        $request->headers->set('Accept', 'application/xml');
         
         $controller->render();
         $expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<response><book><name>How to use PHPUnit</name></book></response>\n";