+++ tests/TestCase/Http/DispatcherTest.php
@@ -19,6 +19,7 @@ use Origin\Http\Response;
 use Origin\Http\Dispatcher;
 
 use Origin\TestSuite\TestTrait;
+use Origin\Core\KeyValueContainer;
 use Origin\Http\Controller\Controller;
 use Origin\Core\Exception\RouterException;
 use Origin\Http\Controller\Exception\MissingMethodException;
@@ -51,7 +52,7 @@ class AnotherMockRequest extends Request
 {
     public function reset()
     {
-        $this->params = [];
+        $this->params = new KeyValueContainer();
     }
 }