+++ tests/TestCase/Http/MiddlewareTest.php
@@ -22,7 +22,7 @@ class MyMiddleware extends Middleware
 {
     public function handle(Request $request): void
     {
-        $request->data('foo', 'bar');
+        $request->data->set('foo', 'bar');
     }
     public function process(Request $request, Response $response): void
     {